bootcamp/week-07-servers_bd_apis_externas/3-proyecto/starter/python-server/pyproject.toml
[project]
name = "library-manager"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "mcp==1.9.0",
    "aiosqlite==0.20.0",
    "httpx==0.27.2",
    "python-dotenv==1.0.1",
]
